correction for windows

+++ src/socket.cpp
@@ -252,7 +252,7 @@ bool blc::network::Socket::writable() const {
 bool blc::network::Socket::waitRead(unsigned int usec) const {
 	fd_set fds;
 	#ifdef __WIN32
-		struct timeval timeout = {0, static_cast<int64_t>(usec)};
+		struct timeval timeout = {0, static_cast<int32_t>(usec)};
 	#else
 		struct timeval timeout = {0, reinterpret_cast<unsigned int>(usec)};
 	#endif
@@ -267,7 +267,7 @@ bool blc::network::Socket::waitRead(unsigned int usec) const {
 bool blc::network::Socket::waitWrite(unsigned int usec) const {
 	fd_set fds;
 	#ifdef __WIN32
-		struct timeval timeout = {0, static_cast<int64_t>(usec)};
+		struct timeval timeout = {0, static_cast<int32_t>(usec)};
 	#else
 		struct timeval timeout = {0, reinterpret_cast<unsigned int>(usec)};
 	#endif
